+++ .travis.yml
@@ -43,6 +43,8 @@ script:
   - cargo build --no-default-features
   - cargo test -v --no-default-features
   - cargo build --features gpu
+  - cargo build --features tesseract
+  - cargo test -v --features tesseract
   - cargo doc --features gpu --no-deps
   - find src/*.rs | xargs rustfmt --write-mode=diff
   # Available until `cargo fmt` landed in stable rust 1.24
